pyproject.toml
[project]
name = "medical-transcription"
version = "0.1.0"
description = "German medical speech-to-text with structured LLM extraction"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.0.0",
    "ollama>=0.1.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "streamlit>=1.31.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "pydub>=0.25.0",
    "soundfile>=0.12.0",
    "gTTS>=2.5.0",
]

[project.scripts]
medical-transcribe = "medical_transcription.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
